pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "korean-sentiment-analysis"
version = "1.0.0"
description = "Production-grade Korean sentiment analysis API using a fine-tuned KcELECTRA transformer model."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Uzbekswe"},
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic>=2.0.0",
]
ui = [
    "streamlit>=1.40.0",
]
dev = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic>=2.0.0",
    "streamlit>=1.40.0",
    "pytest>=7.0",
    "ruff>=0.4.0",
    "httpx>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
